Move ra_ide_api unit tests
